+++ packages/core/package.json
@@ -17,6 +17,7 @@
   "dependencies": {
     "@newrelic/telemetry-sdk": "^0.4.2",
     "@salesforce/core": "^2.24.0",
+    "@salesforce/source-deploy-retrieve": "^5.6.3",
     "@salesforce/ts-sinon": "^1.3.21",
     "adm-zip": "^0.5.6",
     "apex-parser": "^2.11.0",
@@ -32,6 +33,7 @@
     "lodash": "^4.17.21",
     "rimraf": "^3.0.2",
     "semver": "^7.3.5",
+    "sfdc-soup": "^15.0.0",
     "shelljs": "^0.8.4",
     "simple-git": "^2.45.1",
     "tar": "^6.1.9",
