+++ packages/core/src/dependency/Component.ts
@@ -1,0 +1,11 @@
+/**
+ * Component details and package it belongs to
+ */
+export default interface Component {
+  id: string,
+  fullName: string,
+  type: string,
+  file: string,
+  package: string,
+  indexOfPackage: number
+}
