+++ packages/core/src/dependency/DependencyViolation.ts
@@ -1,0 +1,11 @@
+import Component from "./Component";
+
+export default interface DependencyViolation {
+  package: string,
+  indexOfPackage: number,
+  files: string[],
+  fullName: string,
+  type: string,
+  dependency: Component,
+  description: string
+}
