+++ packages/sfpowerscripts-cli/messages/validate.json
@@ -10,5 +10,8 @@
   "deleteScratchOrgFlagDescription": "Delete scratch-org validation environment, after the command has finished running",
   "keysFlagDescription": "Keys to be used while installing any managed package dependencies. Required format is a string of key-value pairs separated by spaces e.g. packageA:pw123 packageB:pw123 packageC:pw123",
   "visualizeChangesAgainstFlagDescription": "Branch to conduct change analysis against. Enables visualization of changes and the components affected",
+  "baseBranchFlagDescription": "The pull request base branch",
+  "enableImpactAnalysisFlagDescription": "Visualize components impacted by changes in pull request",
+  "enableDependencyValidation": "Validate dependencies between packages for changed components",
   "tagFlagDescription":"Tag the build with a label, useful to identify in metrics"
 }
