+++ packages/sfpowerscripts-cli/package.json
@@ -15,7 +15,6 @@
     "@oclif/errors": "^1.3.0",
     "@salesforce/command": "^2",
     "@salesforce/core": "^2.24.0",
-    "@salesforce/source-deploy-retrieve": "^5.6.0",
     "adm-zip": "^0.5.4",
     "ajv": "^8.1.0",
     "async-retry": "^1.3.1",
